complete exercises from 4.15 to 4.22

+++ controllers/blogs.js
@@ -1,10 +1,11 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
 const logger = require("../utils/logger");
+const jwt = require("jsonwebtoken");
 
 blogRouter.get("/", async (request, response) => {
-  logger.info("working");
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user");
   response.json(blogs);
 });
 
@@ -18,23 +19,43 @@ blogRouter.get("/:id", async (request, response) => {
 });
 
 blogRouter.post("/", async (request, response, next) => {
-  logger.info(request.body);
   const body = request.body;
+  console.log(body);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  console.log(decodedToken);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: 0,
+    user: user.id,
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   response.json(savedBlog);
 });
 
 blogRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  const blogUserId = blog.user.toString();
+
+  const user = request.user;
+
+  if (user.id === blogUserId) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).end();
+  }
 });
 
 blogRouter.put("/:id", async (request, response, next) => {
